feat: 핀 상세 - 포스트 / 댓글 제외 연결

+++ app/src/main/java/com/issueissyu/fe/data/remote/dto/request/pin/ApplyPinEmojiRequest.kt
@@ -1,5 +1,5 @@
 package com.issueissyu.fe.data.remote.dto.request.pin
 
 data class ApplyPinEmojiRequest(
-    val emojiId: Int,
+    val emojiId: Long,
 )
